The version="3.0" change was a false alarm so revert that change and replace with a minor clarification.
I got tripped by the instructions referring to the "gtksourceview-3.0 folder" (assuming it meant just "~/.local/share/gtksourceview-3.0") when on closer reading the instructions tell to put the file in ~/.local/share/gtksourceview-3.0/language-specs. This commit clarifies the meaning by adding the "language-spec" (and "styles") subfolder to the GTK3 instructions.